tools/phrasepack_importer/requirements.txt
google-cloud-aiplatform>=1.52.0
